fix: review bugs
