+++ errors/ForbiddenError.js
@@ -1,0 +1,9 @@
+class ForbiddenError extends Error {
+  constructor(message = 'Действие запрещено') {
+    super(message);
+    this.name = 'ForbiddenError';
+    this.statusCode = 403;
+  }
+}
+
+module.exports = ForbiddenError;
