+++ errors/NotFoundError.js
@@ -1,0 +1,9 @@
+class NotFoundError extends Error {
+  constructor(message = 'not found') {
+    super(message);
+    this.name = 'NotFoundError';
+    this.statusCode = 404;
+  }
+}
+
+module.exports = NotFoundError;
