+++ routes/index.js
@@ -4,7 +4,7 @@ const router = require('express').Router();
 const { celebrate, Joi } = require('celebrate');
 const usersRouter = require('./users');
 const moviesRouter = require('./movies');
-const { NotFoundError } = require('../middlewares/errors');
+const NotFoundError = require('../errors/NotFoundError');
 const auth = require('../middlewares/auth');
 const { login, createUser } = require('../controllers/users');
 
@@ -12,7 +12,7 @@ router.post('/signup', celebrate({
   body: Joi.object().keys({
     email: Joi.string().required().email(),
     password: Joi.string().required(),
-    name: Joi.string().min(2).max(30),
+    name: Joi.string().min(2).max(30).required(),
   }),
 }), createUser);
